+++ zoobot/pytorch/training/finetune.py
@@ -870,70 +870,3 @@ def download_from_name(class_name: str, hub_name: str):
         filename=f"{class_name}.ckpt"
     )
     return downloaded_loc
-
-
-
-
-def cosine_schedule(
-    step: int,
-    max_steps: int,
-    start_value: float,
-    end_value: float,
-    period: Optional[int] = None,
-) -> float:
-    """
-    Use cosine decay to gradually modify start_value to reach target end_value during
-    iterations.
-    Copied from lightly library (thank you for open sourcing)
-
-    Args:
-        step:
-            Current step number.
-        max_steps:
-            Total number of steps.
-        start_value:
-            Starting value.
-        end_value:
-            Target value.
-        period (optional):
-            The number of steps over which the cosine function completes a full cycle.
-            If not provided, it defaults to max_steps.
-
-    Returns:
-        Cosine decay value.
-
-    """
-    if step < 0:
-        raise ValueError("Current step number can't be negative")
-    if max_steps < 1:
-        raise ValueError("Total step number must be >= 1")
-    if period is None and step > max_steps:
-        warnings.warn(
-            f"Current step number {step} exceeds max_steps {max_steps}.",
-            category=RuntimeWarning,
-        )
-    if period is not None and period <= 0:
-        raise ValueError("Period must be >= 1")
-
-    decay: float
-    if period is not None:  # "cycle" based on period, if provided
-        decay = (
-            end_value
-            - (end_value - start_value) * (np.cos(2 * np.pi * step / period) + 1) / 2
-        )
-    elif max_steps == 1:
-        # Avoid division by zero
-        decay = end_value
-    elif step == max_steps:
-        # Special case for Pytorch Lightning which updates LR scheduler also for epoch
-        # after last training epoch.
-        decay = end_value
-    else:
-        decay = (
-            end_value
-            - (end_value - start_value)
-            * (np.cos(np.pi * step / (max_steps - 1)) + 1)
-            / 2
-        )
-    return decay
-
